Refactor date input
