+++ dc-cudami-editor/src/components/FormDateInput.jsx
@@ -1,0 +1,45 @@
+import React, {useContext} from 'react'
+import {FormGroup, Label} from 'reactstrap'
+import DatePicker from 'react-date-picker'
+import {useTranslation} from 'react-i18next'
+
+import AppContext from './AppContext'
+
+const formatDate = (date) => {
+  if (date) {
+    // Selected datetime will be displayed correctly in DatePicker.
+    // But after reloading, there's a minus one-day offset. Here
+    // we add 12 hours (could be related to the time zone of the user)
+    // to the user-defined date to workaround this issue 😒
+    date.setHours(date.getHours() + 12)
+    return date.toISOString().slice(0, 10)
+  }
+  return undefined
+}
+
+const FormDateInput = ({id, label, onChange, value}) => {
+  const {uiLocale} = useContext(AppContext)
+  const {t} = useTranslation()
+  return (
+    <FormGroup>
+      <Label className="font-weight-bold" for={id}>
+        {label}
+      </Label>
+      <DatePicker
+        calendarAriaLabel={t('datePicker.toggleCalendar')}
+        clearAriaLabel={t('datePicker.clearDate')}
+        dayAriaLabel={t('datePicker.day')}
+        id={id}
+        locale={uiLocale}
+        monthAriaLabel={t('datePicker.month')}
+        nativeInputAriaLabel={t('datePicker.date')}
+        onChange={(date) => onChange(formatDate(date))}
+        required={false}
+        value={value && new Date(value)}
+        yearAriaLabel={t('datePicker.year')}
+      />
+    </FormGroup>
+  )
+}
+
+export default FormDateInput
