+++ dc-cudami-editor/src/components/WebpageForm.jsx
@@ -13,11 +13,11 @@ import {
 import {useTranslation} from 'react-i18next'
 
 import FormButtons from './FormButtons'
+import FormDateInput from './FormDateInput'
 import FormEditor from './FormEditor'
 import FormInput from './FormInput'
 import LanguageAdder from './LanguageAdder'
 import LanguageTab from './LanguageTab'
-import PublicationDatesForm from './PublicationDatesForm'
 import Teaser from './Teaser'
 
 const WebpageForm = ({
@@ -51,23 +51,38 @@ const WebpageForm = ({
           <hr />
         </Col>
       </Row>
-      <Row>
-        <Col sm="12">
-          {identifiable.uuid && (
+      {identifiable.uuid && (
+        <Row>
+          <Col sm="12">
             <FormInput
               id="uuid"
               label="ID"
               readOnly
               value={identifiable.uuid}
             />
-          )}
-          <PublicationDatesForm
-            onChange={(updateKey, updateValue) =>
-              onUpdate({...identifiable, [updateKey]: updateValue})
-            }
-            publicationEndDate={identifiable.publicationEnd}
-            publicationStartDate={identifiable.publicationStart}
+          </Col>
+        </Row>
+      )}
+      <Row>
+        <Col sm="3">
+          <FormDateInput
+            id="publication-start-date"
+            label={t('publicationStatus.startDate')}
+            onChange={(date) => onUpdate({publicationStart: date})}
+            value={identifiable.publicationStart}
           />
+        </Col>
+        <Col sm="3">
+          <FormDateInput
+            id="publication-end-date"
+            label={t('publicationStatus.endDate')}
+            onChange={(date) => onUpdate({publicationEnd: date})}
+            value={identifiable.publicationEnd}
+          />
+        </Col>
+      </Row>
+      <Row>
+        <Col sm="12">
           <Nav tabs>
             {existingLanguages.map((language) => (
               <LanguageTab
